Remove exposed types
