+++ src/components/body/Body.tsx
@@ -1,7 +1,6 @@
 import React, { useContext, useEffect, useRef, useState } from 'react';
 import PickerContext from '../../context/PickerContext';
 import ProviderContext from '../../context/TenorContext';
-import { TenorImage } from '../../types/exposedTypes';
 import './Body.css';
 import CategoryList from './CategoryList';
 import SearchResult from './SearchResult';
