+++ src/index.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import ErrorBoundary from './components/ErrorBoundary';
 import GifPickerReact, { GifPickerReactProps } from './GifPickerReact';
 
-export { ContentFilter, Theme } from './types/exposedTypes';
-export type { TenorImage } from './types/exposedTypes';
+export { ContentFilter } from './types/TenorTypes';
+export { Theme } from './GifPickerReact';
 
 export interface GifPickerProps extends GifPickerReactProps {}
